🎨 정답 입력 코드를 컴포넌트화
